feat: add react-router boilerplate
