+++ src/App.test.tsx
@@ -1,3 +1,2 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import App from './App';
