--- src/routes/index.tsx
@@ -1,3 +1,0 @@
-// Export all routes here
-
-export {}
