--- src/stores/index.tsx
@@ -1,3 +1,0 @@
-// Export all stores here
-
-export {}
